# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanSphere API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Team {
  id: number;
  name: string;
  code: string;
  flag: string;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
} as const;

export interface Match {
  id: number;
  teamA: Team;
  teamB: Team;
  /** @nullable */
  scoreA?: number | null;
  /** @nullable */
  scoreB?: number | null;
  datetime: string;
  status: MatchStatus;
  group: string;
  /** @nullable */
  minute?: number | null;
  /** @nullable */
  stadium?: string | null;
}

export type PredictionStatus = typeof PredictionStatus[keyof typeof PredictionStatus];


export const PredictionStatus = {
  pending: 'pending',
  correct: 'correct',
  partial: 'partial',
  wrong: 'wrong',
} as const;

export interface Prediction {
  id: number;
  matchId: number;
  teamA: Team;
  teamB: Team;
  predictedScoreA: number;
  predictedScoreB: number;
  /** @nullable */
  pointsEarned?: number | null;
  status: PredictionStatus;
}

export interface LeagueSummary {
  id: number;
  name: string;
  /** @nullable */
  emoji?: string | null;
  myRank: number;
  myPoints: number;
  memberCount: number;
}

export interface Dashboard {
  featuredMatch: Match;
  todayMatches: Match[];
  myPredictions: Prediction[];
  myLeague: LeagueSummary;
}

export type MatchEventType = typeof MatchEventType[keyof typeof MatchEventType];


export const MatchEventType = {
  goal: 'goal',
  yellow_card: 'yellow_card',
  red_card: 'red_card',
  substitution: 'substitution',
} as const;

export interface MatchEvent {
  id: number;
  type: MatchEventType;
  minute: number;
  team: string;
  player: string;
}

export interface MatchStat {
  label: string;
  valueA: string;
  valueB: string;
}

export interface MatchDetail {
  match: Match;
  events: MatchEvent[];
  stats: MatchStat[];
  userPrediction?: Prediction;
}

export type TeamStandingStatus = typeof TeamStandingStatus[keyof typeof TeamStandingStatus];


export const TeamStandingStatus = {
  qualified: 'qualified',
  danger: 'danger',
  eliminated: 'eliminated',
  pending: 'pending',
} as const;

export interface TeamStanding {
  team: Team;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDiff?: number;
  points: number;
  status: TeamStandingStatus;
}

export interface GroupStanding {
  group: string;
  teams: TeamStanding[];
  /** @nullable */
  keyScenario: string | null;
}

export interface PredictionInput {
  matchId: number;
  /** @minimum 0 */
  predictedScoreA: number;
  /** @minimum 0 */
  predictedScoreB: number;
}

export interface LeagueMember {
  rank: number;
  name: string;
  initials: string;
  points: number;
  change: number;
  isCurrentUser?: boolean;
}

export interface League {
  id: number;
  name: string;
  /** @nullable */
  emoji?: string | null;
  memberCount: number;
  myRank: number;
  myPoints: number;
  inviteCode: string;
}

export interface LeagueDetail {
  id: number;
  name: string;
  /** @nullable */
  emoji?: string | null;
  memberCount: number;
  myRank: number;
  myPoints: number;
  inviteCode: string;
  members: LeagueMember[];
  badges: string[];
}

export interface Profile {
  id: number;
  name: string;
  initials: string;
  fanSince: number;
  location: string;
  totalPoints: number;
  predictionAccuracy: number;
  predictionCount: number;
  favoriteTeams: Team[];
}

export type GetMatchesParams = {
date?: string;
status?: string;
};

export type GetStandingsParams = {
group?: string;
};