# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FanSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: List all matches
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, code, flag]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        flag:
          type: string
    Match:
      type: object
      required: [id, teamA, teamB, datetime, status, group]
      properties:
        id:
          type: integer
        teamA:
          $ref: "#/components/schemas/Team"
        teamB:
          $ref: "#/components/schemas/Team"
        scoreA:
          type: ["integer", "null"]
        scoreB:
          type: ["integer", "null"]
        datetime:
          type: string
        status:
          type: string
          enum: [upcoming, live, finished]
        group:
          type: string
        minute:
          type: ["integer", "null"]
        stadium:
          type: ["string", "null"]
    MatchEvent:
      type: object
      required: [id, type, minute, team, player]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [goal, yellow_card, red_card, substitution]
        minute:
          type: integer
        team:
          type: string
        player:
          type: string
    MatchStat:
      type: object
      required: [label, valueA, valueB]
      properties:
        label:
          type: string
        valueA:
          type: string
        valueB:
          type: string
    MatchDetail:
      type: object
      required: [match, events, stats]
      properties:
        match:
          $ref: "#/components/schemas/Match"
        events:
          type: array
          items:
            $ref: "#/components/schemas/MatchEvent"
        stats:
          type: array
          items:
            $ref: "#/components/schemas/MatchStat"
        userPrediction:
          $ref: "#/components/schemas/Prediction"
...
    Prediction:
      type: object
      required: [id, matchId, teamA, teamB, predictedScoreA, predictedScoreB, status]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamA:
          $ref: "#/components/schemas/Team"
        teamB:
          $ref: "#/components/schemas/Team"
        predictedScoreA:
          type: integer
        predictedScoreB:
          type: integer
        pointsEarned:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, correct, partial, wrong]
...