# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanSphere API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatus } from './matchStatus';
import type { Team } from './team';

export interface Match {
  id: number;
  teamA: Team;
  teamB: Team;
  /** @nullable */
  scoreA?: number | null;
  /** @nullable */
  scoreB?: number | null;
  datetime: string;
  status: MatchStatus;
  group: string;
  /** @nullable */
  minute?: number | null;
  /** @nullable */
  stadium?: string | null;
}