# lib/api-zod/src/generated/types/matchEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanSphere API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchEventType } from './matchEventType';

export interface MatchEvent {
  id: number;
  type: MatchEventType;
  minute: number;
  team: string;
  player: string;
}